test(shared-server): finish client state test ownership
