+++ packages/tests/shared-server/client-state/app-client-inbox-operation-matrix.test.ts
@@ -1,79 +1,37 @@
-import { describe, expect, expectTypeOf, it, vi } from 'vitest';
+import { describe, expect, it, vi } from 'vitest';
 
-import type { InboxQueueReader } from '@shared/services/InboxQueueReader.ts';
-import * as sharedServer from '@shared-server/mod.ts';
-// prettier-ignore
-import * as compatibilityInbox from
-  '@shared-server/rallar-system/services/AppClientInboxService.ts';
-// prettier-ignore
-import * as compatibilityService from
-  '@shared-server/rallar-system/services/client-state-service.ts';
-// prettier-ignore
-import {
-  ClientMutationIdempotencyConflictError,
-} from
-'@shared-server/rallar-system/client-state/mutation/result-validation/validate-client-mutation.ts';
-// prettier-ignore
-import {
-  ClientMutationRejectedError,
-} from '@shared-server/rallar-system/client-state/client-state-validation-primitives.ts';
-import {
-  toClientMutationIssuedSessionAuthority,
-  toClientMutationSystemAuthority,
-} from '@shared-server/rallar-system/client-state/mutation/client-mutation-authority.ts';
-import {
-  toClientMutationCommand,
-  toConnectCommandInput,
-  toDisconnectCommandInput,
-  toExpiryCommandInput,
-  toHeartbeatCommandInput,
-  toUpsertInstanceCommandInput,
-  toUpsertPrincipalCommandInput,
-} from '@shared-server/rallar-system/client-state/mutation/client-mutation-command.ts';
-// prettier-ignore
-import {
-  createClientStateService,
-} from '@shared-server/rallar-system/client-state/client-state-service.ts';
-import {
-  requiresClientWrite,
-  toClientMutationReceipt,
-  toClientStateWritten,
-} from '@shared-server/rallar-system/client-state/client-state-service-contracts.ts';
-// prettier-ignore
-import {
-  AppClientInboxService,
-} from '@shared-server/rallar-system/client-state/inbox/app-client-inbox-service.ts';
-import {
-  AppInboxService,
-  AppInboxType,
-} from '@shared-server/rallar-system/services/AppInboxService.ts';
-
-import type * as PublicSharedServer from '@shared-server/mod.ts';
+import type { ClientSnapshot } from '@shared/api/client-types.ts';
+import { NonRetryableException } from '@shared/queuebox/DequeueResourceEntryController.ts';
+import { InboxQueueReader } from '@shared/services/InboxQueueReader.ts';
 import type {
-  ClientAuthorisedWsSessionConnectAppInboxPayload,
-  ClientAuthorisedWsSessionDisconnectAppInboxPayload,
-  ClientExpiredSessionsAppInboxPayload,
   ClientInstanceUpsertAppInboxPayload,
   ClientPrincipalUpsertAppInboxPayload,
   ClientSessionConnectAppInboxPayload,
   ClientSessionDisconnectAppInboxPayload,
   ClientSessionHeartbeatAppInboxPayload,
 } from '@shared-server/rallar-system/client-state/inbox/app-client-inbox-contracts.ts';
-// prettier-ignore
+import { AppClientInboxService } from '@shared-server/rallar-system/client-state/inbox/app-client-inbox-service.ts';
 import type {
-  ClientMutationPersistedFacts,
-} from '@shared-server/rallar-system/client-state/mutation/client-mutation-command.ts';
-// prettier-ignore
-import type {
-  ClientMutationReceipt,
-} from '@shared-server/rallar-system/client-state/mutation/client-mutation-contracts.ts';
-import type {
-  ClientMutationWritten,
   ClientStateService,
-  ClientStateServiceDependencies,
   ClientStateWritten,
-  RegisterAuthorisedWsClientInput,
 } from '@shared-server/rallar-system/client-state/client-state-service-contracts.ts';
+import {
+  AppInboxService,
+  AppInboxType,
+} from '@shared-server/rallar-system/services/AppInboxService.ts';
+
+import { createAppInboxTestDatabase } from '../app-inbox-test-database.ts';
+import { FakeRuntimeStateRepository } from '../fake-runtime-state-repository.ts';
+import {
+  CLIENT_STATE_TEST_SCOPE as SCOPE,
+  TestResourceInbox,
+  TestResourceInboxResults,
+  createAutoAuthorizingClientStateService,
+  createClientStateServiceStub,
+  createPublisher,
+  processAppInbox,
+  requireRightSnapshot,
+} from './app-client-inbox-mutation-test-harness.ts';
 
 describe('AppClientInbox operation matrix', () => {
   it('registers the established eight client mutation families in order', () => {
@@ -99,122 +57,6 @@ describe('AppClientInbox operation matrix', () => {
   });
 });
 
-describe('shared-server client-state public API contract', () => {
-  it(
-    'keeps every predecessor runtime export as an identical canonical binding',
-    assertPredecessorRuntimeExports,
-  );
-  it(
-    'keeps public, canonical, and compatibility runtime exports identical',
-    assertPublicCanonicalCompatibilityRuntimeIdentity,
-  );
-  it('does not leak internal client-state construction or handler bindings', assertNoInternalLeaks);
-  it(
-    'keeps predecessor public types equal to their canonical and compatibility contracts',
-    assertPredecessorPublicTypeIdentity,
-  );
-});
-
-// @ts-expect-error Public callers must not receive the handler-only mutation capability.
-type PublicClientStateMutationService = PublicSharedServer.ClientStateMutationService;
-// @ts-expect-error Public callers must not receive the construction-only service factory type.
-type PublicClientStateServiceFactory = PublicSharedServer.ClientStateServiceFactory;
-// @ts-expect-error Public callers must not receive the timing factory type.
-type PublicClientStateServiceTimingFactory = PublicSharedServer.ClientStateServiceTimingFactory;
-type PublicClientStateInboxHandlerDependencies =
-  // @ts-expect-error Public callers must not receive handler dependency wiring.
-  PublicSharedServer.ClientStateInboxHandlerDependencies;
-// @ts-expect-error Public callers must not receive handler private after-commit data.
-type PublicClientStateInboxAfterCommitResult = PublicSharedServer.ClientStateInboxAfterCommitResult;
-
-function assertPredecessorRuntimeExports(): void {
-  expect(sharedServer.ClientMutationIdempotencyConflictError).toBe(
-    ClientMutationIdempotencyConflictError,
-  );
-  expect(sharedServer.ClientMutationRejectedError).toBe(ClientMutationRejectedError);
-  expect(sharedServer.toClientMutationIssuedSessionAuthority).toBe(
-    toClientMutationIssuedSessionAuthority,
-  );
-  expect(sharedServer.toClientMutationSystemAuthority).toBe(toClientMutationSystemAuthority);
-  expect(sharedServer.toClientMutationCommand).toBe(toClientMutationCommand);
-  expect(sharedServer.toConnectCommandInput).toBe(toConnectCommandInput);
-  expect(sharedServer.toDisconnectCommandInput).toBe(toDisconnectCommandInput);
-  expect(sharedServer.toExpiryCommandInput).toBe(toExpiryCommandInput);
-  expect(sharedServer.toHeartbeatCommandInput).toBe(toHeartbeatCommandInput);
-  expect(sharedServer.toUpsertInstanceCommandInput).toBe(toUpsertInstanceCommandInput);
-  expect(sharedServer.toUpsertPrincipalCommandInput).toBe(toUpsertPrincipalCommandInput);
-  expect(sharedServer.createClientStateService).toBe(createClientStateService);
-  expect(sharedServer.requiresClientWrite).toBe(requiresClientWrite);
-  expect(sharedServer.toClientMutationReceipt).toBe(toClientMutationReceipt);
-  expect(sharedServer.toClientStateWritten).toBe(toClientStateWritten);
-  expect(sharedServer.AppClientInboxService).toBe(AppClientInboxService);
-}
-
-function assertPublicCanonicalCompatibilityRuntimeIdentity(): void {
-  expect(sharedServer.createClientStateService).toBe(compatibilityService.createClientStateService);
-  expect(sharedServer.AppClientInboxService).toBe(compatibilityInbox.AppClientInboxService);
-  expect(sharedServer.ClientMutationRejectedError).toBe(
-    compatibilityService.ClientMutationRejectedError,
-  );
-  expect(sharedServer.ClientMutationIdempotencyConflictError).toBe(
-    compatibilityService.ClientMutationIdempotencyConflictError,
-  );
-  expect(sharedServer.toClientMutationCommand).toBe(compatibilityService.toClientMutationCommand);
-  expect(sharedServer.toClientMutationIssuedSessionAuthority).toBe(
-    compatibilityService.toClientMutationIssuedSessionAuthority,
-  );
-}
-
-function assertNoInternalLeaks(): void {
-  for (const internalExport of [
-    'CLIENT_STATE_INBOX_REGISTRATION_TYPES',
-    'createTimedClientStateService',
-    'ClientStateInboxHandler',
-  ]) {
-    expect(Object.hasOwn(sharedServer, internalExport), internalExport).toBe(false);
-  }
-}
-
-function assertPredecessorPublicTypeIdentity(): void {
-  // prettier-ignore
-  expectTypeOf<PublicSharedServer.ClientMutationPersistedFacts>()
-    .toEqualTypeOf<ClientMutationPersistedFacts>();
-  expectTypeOf<PublicSharedServer.ClientMutationReceipt>().toEqualTypeOf<ClientMutationReceipt>();
-  // prettier-ignore
-  expectTypeOf<PublicSharedServer.RegisterAuthorisedWsClientInput>()
-    .toEqualTypeOf<RegisterAuthorisedWsClientInput>();
-  expectTypeOf<PublicSharedServer.ClientMutationWritten>().toEqualTypeOf<ClientMutationWritten>();
-  expectTypeOf<PublicSharedServer.ClientStateWritten>().toEqualTypeOf<ClientStateWritten>();
-  expectTypeOf<PublicSharedServer.ClientStateService>().toEqualTypeOf<ClientStateService>();
-  // prettier-ignore
-  expectTypeOf<PublicSharedServer.ClientStateServiceDependencies>()
-    .toEqualTypeOf<ClientStateServiceDependencies>();
-  // prettier-ignore
-  expectTypeOf<PublicSharedServer.ClientPrincipalUpsertAppInboxPayload>()
-    .toEqualTypeOf<ClientPrincipalUpsertAppInboxPayload>();
-  // prettier-ignore
-  expectTypeOf<PublicSharedServer.ClientInstanceUpsertAppInboxPayload>()
-    .toEqualTypeOf<ClientInstanceUpsertAppInboxPayload>();
-  // prettier-ignore
-  expectTypeOf<PublicSharedServer.ClientSessionConnectAppInboxPayload>()
-    .toEqualTypeOf<ClientSessionConnectAppInboxPayload>();
-  // prettier-ignore
-  expectTypeOf<PublicSharedServer.ClientSessionHeartbeatAppInboxPayload>()
-    .toEqualTypeOf<ClientSessionHeartbeatAppInboxPayload>();
-  // prettier-ignore
-  expectTypeOf<PublicSharedServer.ClientSessionDisconnectAppInboxPayload>()
-    .toEqualTypeOf<ClientSessionDisconnectAppInboxPayload>();
-  // prettier-ignore
-  expectTypeOf<PublicSharedServer.ClientAuthorisedWsSessionConnectAppInboxPayload>()
-    .toEqualTypeOf<ClientAuthorisedWsSessionConnectAppInboxPayload>();
-  // prettier-ignore
-  expectTypeOf<PublicSharedServer.ClientAuthorisedWsSessionDisconnectAppInboxPayload>()
-    .toEqualTypeOf<ClientAuthorisedWsSessionDisconnectAppInboxPayload>();
-  // prettier-ignore
-  expectTypeOf<PublicSharedServer.ClientExpiredSessionsAppInboxPayload>()
-    .toEqualTypeOf<ClientExpiredSessionsAppInboxPayload>();
-}
-
 function createClientInboxServiceForRegistration(): AppClientInboxService {
   const registrationService = {
     sessionGenerationLifecycle: {} as ClientStateService['sessionGenerationLifecycle'],
@@ -227,4 +69,245 @@ function createClientInboxServiceForRegistration(): AppClientInboxService {
     registrationService as ClientStateService,
     'client-registration-service',
   );
+}
+
+describe('AppClientInbox mutation processing', () => {
+  it('processes principal, instance, and session mutations through the inbox', async () => {
+    const harness = createMutationProcessingHarness();
+    const principal = await upsertPrincipal(harness);
+    const instance = await upsertInstance(harness);
+    const connected = await connectSession(harness);
+    const heartbeat = await heartbeatSession(harness);
+    const disconnected = await disconnectSession(harness);
+
+    expectMutationProcessingResults(harness, {
+      principal,
+      instance,
+      connected,
+      heartbeat,
+      disconnected,
+    });
+  });
+
+  it('returns a left result when a client inbox mutation handler fails with a non-retryable error', async () => {
+    const queue = new TestResourceInbox();
+    const reader = new InboxQueueReader(queue);
+    const results = new TestResourceInboxResults();
+    const service = new AppClientInboxService(
+      reader,
+      queue as never,
+      results as never,
+      createAppInboxTestDatabase(queue, results),
+      createClientStateServiceStub({
+        read: vi.fn(async () => {
+          throw new NonRetryableException('Client principal update failed');
+        }),
+      }),
+      'server-12345678',
+    );
+
+    const result = await processAppInbox<ClientPrincipalUpsertAppInboxPayload, ClientSnapshot>(
+      service,
+      reader,
+      {
+        type: AppInboxType.CLIENT_PRINCIPAL_UPSERT,
+        resourceId: 'upsert-client-fail',
+        contextId: `${SCOPE.applicationId}:${SCOPE.workspaceId}:alice`,
+        senderId: 'alice',
+        data: {
+          scope: SCOPE,
+          principalId: 'alice',
+          request: {
+            username: 'alice',
+            actorPrincipalId: 'alice',
+            requestId: 'upsert-client-fail',
+          },
+        },
+      },
+    );
+
+    expect(result.left).toBe('Client principal update failed');
+  });
+});
+
+function createMutationProcessingHarness() {
+  const queue = new TestResourceInbox();
+  const reader = new InboxQueueReader(queue);
+  const results = new TestResourceInboxResults();
+  const publisher = createPublisher();
+  const runtimeRepository = new FakeRuntimeStateRepository();
+  const database = createAppInboxTestDatabase(queue, results, { runtimeRepository });
+  return {
+    connectedAtEpochMs: Date.now(),
+    publisher,
+    reader,
+    service: new AppClientInboxService(
+      reader,
+      queue as never,
+      results as never,
+      database,
+      createAutoAuthorizingClientStateService(runtimeRepository, database),
+      'server-12345678',
+    ),
+  };
+}
+
+type MutationProcessingHarness = ReturnType<typeof createMutationProcessingHarness>;
+
+function upsertPrincipal(harness: MutationProcessingHarness) {
+  return processAppInbox<ClientPrincipalUpsertAppInboxPayload, ClientStateWritten>(
+    harness.service,
+    harness.reader,
+    {
+      type: AppInboxType.CLIENT_PRINCIPAL_UPSERT,
+      resourceId: 'upsert-client-alice',
+      contextId: `${SCOPE.applicationId}:${SCOPE.workspaceId}:alice`,
+      senderId: 'alice',
+      data: {
+        scope: SCOPE,
+        principalId: 'alice',
+        request: {
+          username: 'alice',
+          displayName: 'Alice',
+          actorPrincipalId: 'alice',
+          requestId: 'upsert-client-alice',
+        },
+      },
+    },
+  );
+}
+
+function upsertInstance(harness: MutationProcessingHarness) {
+  return processAppInbox<ClientInstanceUpsertAppInboxPayload, ClientStateWritten>(
+    harness.service,
+    harness.reader,
+    {
+      type: AppInboxType.CLIENT_INSTANCE_UPSERT,
+      resourceId: 'upsert-client-alice-instance',
+      contextId: `${SCOPE.applicationId}:${SCOPE.workspaceId}:alice`,
+      senderId: 'alice',
+      data: {
+        scope: SCOPE,
+        principalId: 'alice',
+        clientInstanceId: 'alice-browser',
+        request: {
+          platform: 'web',
+          capabilities: ['ws'],
+          actorPrincipalId: 'alice',
+          requestId: 'upsert-client-alice-instance',
+        },
+      },
+    },
+  );
+}
+
+function connectSession(harness: MutationProcessingHarness) {
+  return processAppInbox<ClientSessionConnectAppInboxPayload, ClientStateWritten>(
+    harness.service,
+    harness.reader,
+    {
+      type: AppInboxType.CLIENT_SESSION_CONNECT,
+      resourceId: 'connect-client-alice-session',
+      contextId: `${SCOPE.applicationId}:${SCOPE.workspaceId}:alice`,
+      senderId: 'alice',
+      data: {
+        scope: SCOPE,
+        principalId: 'alice',
+        clientInstanceId: 'alice-browser',
+        sessionId: 'alice-session',
+        request: {
+          generationId: 'generation-alice-session',
+          presenceState: 'online',
+          actorPrincipalId: 'alice',
+          actorSessionId: 'alice-session',
+          connectedAtEpochMs: harness.connectedAtEpochMs,
+          lastHeartbeatAtEpochMs: harness.connectedAtEpochMs,
+          expiresAtEpochMs: harness.connectedAtEpochMs + 60_000,
+          requestId: 'connect-client-alice-session',
+        },
+      },
+    },
+  );
+}
+
+function heartbeatSession(harness: MutationProcessingHarness) {
+  return processAppInbox<ClientSessionHeartbeatAppInboxPayload, ClientStateWritten>(
+    harness.service,
+    harness.reader,
+    {
+      type: AppInboxType.CLIENT_SESSION_HEARTBEAT,
+      resourceId: 'heartbeat-client-alice-session',
+      contextId: `${SCOPE.applicationId}:${SCOPE.workspaceId}:alice`,
+      senderId: 'alice',
+      data: {
+        scope: SCOPE,
+        principalId: 'alice',
+        clientInstanceId: 'alice-browser',
+        sessionId: 'alice-session',
+        request: {
+          generationId: 'generation-alice-session',
+          presenceState: 'away',
+          actorPrincipalId: 'alice',
+          actorSessionId: 'alice-session',
+          lastHeartbeatAtEpochMs: harness.connectedAtEpochMs + 1,
+          expiresAtEpochMs: harness.connectedAtEpochMs + 60_001,
+          requestId: 'heartbeat-client-alice-session',
+        },
+      },
+    },
+  );
+}
+
+function disconnectSession(harness: MutationProcessingHarness) {
+  return processAppInbox<ClientSessionDisconnectAppInboxPayload, ClientStateWritten>(
+    harness.service,
+    harness.reader,
+    {
+      type: AppInboxType.CLIENT_SESSION_DISCONNECT,
+      resourceId: 'disconnect-client-alice-session',
+      contextId: `${SCOPE.applicationId}:${SCOPE.workspaceId}:alice`,
+      senderId: 'alice',
+      data: {
+        scope: SCOPE,
+        principalId: 'alice',
+        clientInstanceId: 'alice-browser',
+        sessionId: 'alice-session',
+        request: {
+          generationId: 'generation-alice-session',
+          reason: 'closed',
+          actorPrincipalId: 'alice',
+          actorSessionId: 'alice-session',
+          requestId: 'disconnect-client-alice-session',
+        },
+      },
+    },
+  );
+}
+
+interface MutationProcessingResults {
+  readonly principal: ClientStateWritten;
+  readonly instance: ClientStateWritten;
+  readonly connected: ClientStateWritten;
+  readonly heartbeat: ClientStateWritten;
+  readonly disconnected: ClientStateWritten;
+}
+
+function expectMutationProcessingResults(
+  harness: MutationProcessingHarness,
+  results: MutationProcessingResults,
+): void {
+  expect(requireRightSnapshot(results.principal).principal.displayName).toBe('Alice');
+  expect(requireRightSnapshot(results.instance).instances[0]).toMatchObject({
+    clientInstanceId: 'alice-browser',
+    platform: 'web',
+  });
+  expect(requireRightSnapshot(results.connected).activeSessions).toHaveLength(1);
+  expect(requireRightSnapshot(results.heartbeat).activeSessions[0]).toMatchObject({
+    sessionId: 'alice-session',
+    presenceState: 'away',
+    lastHeartbeatAtEpochMs: harness.connectedAtEpochMs + 1,
+  });
+  expect(requireRightSnapshot(results.disconnected).activeSessions).toHaveLength(0);
+  expect(harness.publisher.publishClientSnapshot).not.toHaveBeenCalled();
+  expect(harness.publisher.publishClientEvent).not.toHaveBeenCalled();
 }
