+++ packages/tests/shared-server/client-state/client-mutation-command-and-request.test.ts
@@ -1,6 +1,17 @@
-import { describe, expect, it } from 'vitest';
+import { describe, expect, expectTypeOf, it } from 'vitest';
 
+import type { ClientSession } from '@shared/api/client-types.ts';
+import type { ConnectClientSessionRequest } from '@shared/api/state-types.ts';
 import type { StateScope } from '@shared/api/state-types.ts';
+import { ClientStateRepository } from '@shared-server/rallar-system/client-state/persistence/client-state-repository.ts';
+import {
+  clientStateInstanceStorageKey,
+  clientStatePrincipalStorageKey,
+  clientStateSessionStorageKey,
+  decodeClientPrincipalStorageKey,
+} from '@shared-server/rallar-system/client-state/persistence/client-state-storage-keys.ts';
+import { clientStatePrincipalStorageKey as compatibilityClientStatePrincipalStorageKey } from '@shared-server/rallar-system/client-state-storage-keys.ts';
+import { ClientStateRepository as compatibilityClientStateRepository } from '@shared-server/rallar-system/repositories/ClientStateRepository.ts';
 import {
   toClientMutationCommand,
   toConnectCommandInput,
@@ -16,6 +27,16 @@ import {
   toClientMutationCommand as legacyToClientMutationCommand,
   toConnectCommandInput as legacyToConnectCommandInput,
 } from '@shared-server/rallar-system/services/client-state-service.ts';
+import {
+  ClientMutationRejectedError,
+  validateClientMutationCommand,
+} from '@shared-server/rallar-system/services/client-state-mutations.ts';
+
+import {
+  clientMutationPrincipalRef as principalRef,
+  invalidSessionCommand,
+  validFacts,
+} from './client-mutation-validation-test-fixtures.ts';
 
 const scope: StateScope = {
   applicationId: 'app-1',
@@ -133,3 +154,127 @@ describe('client mutation command and request projection', () => {
     expect(legacyToConnectCommandInput).toBe(toConnectCommandInput);
   });
 });
+
+describe('client mutation command and compatibility contracts', () => {
+  it('keeps canonical encoded keys and compatibility repository identities stable', () => {
+    const principal = {
+      applicationId: 'app:/%',
+      workspaceId: '_',
+      principalId: 'alice smith',
+    };
+    const principalKey = clientStatePrincipalStorageKey(principal);
+
+    expect(principalKey).toBe('app=app%3A%2F%25:ws=_:principal=alice%20smith');
+    expect(decodeClientPrincipalStorageKey(principalKey)).toEqual(principal);
+    expect(compatibilityClientStatePrincipalStorageKey(principal)).toBe(principalKey);
+    expect(clientStateInstanceStorageKey({ ...principal, clientInstanceId: 'web/1' })).toBe(
+      `${principalKey}:instance=web%2F1`,
+    );
+    expect(
+      clientStateSessionStorageKey({
+        ...principal,
+        clientInstanceId: 'web/1',
+        sessionId: 'session:1',
+      }),
+    ).toBe(`${principalKey}:instance=web%2F1:session=session%3A1`);
+    expect(compatibilityClientStateRepository).toBe(ClientStateRepository);
+  });
+
+  it('requires generation identity and exposes no caller command hash', () => {
+    expectTypeOf<ConnectClientSessionRequest>().toHaveProperty('generationId');
+    expectTypeOf<ConnectClientSessionRequest>().not.toHaveProperty('commandHash');
+    expectTypeOf<ClientSession>().toHaveProperty('generationVersion');
+  });
+
+  it('rejects malformed authoritative command shapes for every client branch before compute or hash', () => {
+    for (const command of malformedAuthoritativeCommands()) {
+      expect(() => validateClientMutationCommand(command)).toThrow(ClientMutationRejectedError);
+    }
+  });
+});
+
+const malformedCommandBase = {
+  aggregateRef: principalRef('alice'),
+  commandId: 'command-1',
+  requestId: 'command-1',
+  facts: validFacts(),
+} as const;
+const malformedCommandActor = {
+  actorPrincipalId: null,
+  actorSessionId: null,
+  reason: null,
+  traceId: null,
+} as const;
+
+function malformedAuthoritativeCommands(): readonly unknown[] {
+  return [
+    [],
+    malformedPrincipalCommand(),
+    malformedInstanceCommand(),
+    invalidSessionCommand(malformedCommandBase, malformedCommandActor, 'connectSession', {
+      generationId: { forged: true },
+    }),
+    invalidSessionCommand(
+      malformedCommandBase,
+      malformedCommandActor,
+      'connectAuthorisedWsSession',
+      {
+        transport: 'carrier-pigeon',
+      },
+    ),
+    invalidSessionCommand(malformedCommandBase, malformedCommandActor, 'heartbeatSession', {
+      lastHeartbeatAtEpochMs: -1,
+    }),
+    invalidSessionCommand(malformedCommandBase, malformedCommandActor, 'disconnectSession', {
+      actorPrincipalId: 42,
+    }),
+    invalidSessionCommand(
+      malformedCommandBase,
+      malformedCommandActor,
+      'disconnectAuthorisedWsSession',
+      {
+        reason: { nested: 'not-a-string' },
+      },
+    ),
+    invalidSessionCommand(malformedCommandBase, malformedCommandActor, 'expireSession', {
+      generationVersion: 0,
+      observedExpiresAtEpochMs: Number.NaN,
+    }),
+  ];
+}
+
+function malformedPrincipalCommand() {
+  return {
+    ...malformedCommandBase,
+    operation: 'upsertPrincipal',
+    input: {
+      ...malformedCommandActor,
+      username: '',
+      displayName: null,
+      avatarUrl: null,
+      status: 'impossible',
+      authProvider: null,
+      externalSubjectId: null,
+      roles: [],
+      metadata: {},
+      lastSeenAtEpochMs: null,
+    },
+  };
+}
+
+function malformedInstanceCommand() {
+  return {
+    ...malformedCommandBase,
+    operation: 'upsertInstance',
+    clientInstanceId: '',
+    input: {
+      ...malformedCommandActor,
+      status: null,
+      platform: 'browser',
+      deviceLabel: null,
+      appVersion: null,
+      userAgent: null,
+      capabilities: [],
+    },
+  };
+}
