+++ packages/tests/shared-server/client-state/client-mutation-idempotency.test.ts
@@ -1,6 +1,26 @@
 import { describe, expect, it } from 'vitest';
 
+import { InboxQueueReader } from '@shared/services/InboxQueueReader.ts';
+import type { ClientPrincipalUpsertAppInboxPayload } from '@shared-server/rallar-system/client-state/inbox/app-client-inbox-contracts.ts';
+import { AppClientInboxService } from '@shared-server/rallar-system/client-state/inbox/app-client-inbox-service.ts';
 import { computeClientMutation } from '@shared-server/rallar-system/client-state/mutation/compute/compute-client-mutation.ts';
+import { ClientStateRepository } from '@shared-server/rallar-system/client-state/persistence/client-state-repository.ts';
+import type { ClientStateWritten } from '@shared-server/rallar-system/client-state/client-state-service-contracts.ts';
+import { AppInboxType } from '@shared-server/rallar-system/services/AppInboxService.ts';
+import { ClientMutationIdempotencyConflictError } from '@shared-server/rallar-system/services/client-state-service.ts';
+
+import { createAppInboxTestDatabase } from '../app-inbox-test-database.ts';
+import { FakeRuntimeStateRepository } from '../fake-runtime-state-repository.ts';
+import {
+  CLIENT_STATE_TEST_SCOPE as APP_SCOPE,
+  TestResourceInbox,
+  TestResourceInboxResults,
+  createAutoAuthorizingClientStateService,
+  createPublisher,
+  processAppInbox,
+  requireRightSnapshot,
+  requireRightWritten,
+} from './app-client-inbox-mutation-test-harness.ts';
 
 import {
   emptyRead,
@@ -9,6 +29,21 @@ import {
   readAfterWrite,
   requireWrite,
 } from './client-mutation-compute-test-fixtures.ts';
+import {
+  AggregateBarrierRepository,
+  createService,
+  outboxFor,
+} from './client-mutation-concurrency-test-runtime.ts';
+import {
+  CLIENT_MUTATION_SERVICE_SCOPE,
+  createPublisher as createServicePublisher,
+  toClientPrincipalRef,
+} from './client-state-service-test-fixtures.ts';
+import {
+  CLIENT_MUTATION_TEST_SCOPE as SCOPE,
+  clientMutationPrincipalRef as principalRef,
+} from './client-mutation-validation-test-fixtures.ts';
+import { createLegacyClientStateTestDriver as createClientStateService } from './client-state-test-runtime.ts';
 
 describe('client mutation idempotency compute', () => {
   it('replays the exact stored receipt, snapshot, and event', async () => {
@@ -46,5 +81,209 @@ describe('client mutation idempotency compute', () => {
       existingCommandHash: command.facts.commandHash,
       receivedCommandHash: conflicting.facts.commandHash,
     });
+  });
+});
+
+describe('client mutation AppInbox idempotency', () => {
+  it('replays stored idempotent mutation results without direct publication', async () => {
+    const { publisher, reader, service } = createAppInboxIdempotencyHarness();
+
+    const first = await processAppInbox<ClientPrincipalUpsertAppInboxPayload, ClientStateWritten>(
+      service,
+      reader,
+      {
+        type: AppInboxType.CLIENT_PRINCIPAL_UPSERT,
+        resourceId: 'upsert-client-alice-first',
+        contextId: `${APP_SCOPE.applicationId}:${APP_SCOPE.workspaceId}:alice`,
+        senderId: 'alice',
+        data: {
+          scope: APP_SCOPE,
+          principalId: 'alice',
+          request: {
+            username: 'alice',
+            displayName: 'Alice',
+            actorPrincipalId: 'alice',
+            requestId: 'upsert-client-alice',
+          },
+        },
+      },
+    );
+    vi.mocked(publisher.publishClientSnapshot).mockClear();
+    vi.mocked(publisher.publishClientEvent).mockClear();
+
+    const replay = await processAppInbox<ClientPrincipalUpsertAppInboxPayload, ClientStateWritten>(
+      service,
+      reader,
+      {
+        type: AppInboxType.CLIENT_PRINCIPAL_UPSERT,
+        resourceId: 'upsert-client-alice-replay',
+        contextId: `${APP_SCOPE.applicationId}:${APP_SCOPE.workspaceId}:alice`,
+        senderId: 'alice',
+        data: {
+          scope: APP_SCOPE,
+          principalId: 'alice',
+          request: {
+            username: 'alice',
+            displayName: 'Alice',
+            actorPrincipalId: 'alice',
+            requestId: 'upsert-client-alice',
+          },
+        },
+      },
+    );
+
+    expect(requireRightSnapshot(replay).principal.displayName).toBe('Alice');
+    expect(requireRightWritten(replay).event).toEqual(requireRightWritten(first).event);
+    expect(publisher.publishClientSnapshot).not.toHaveBeenCalled();
+    expect(publisher.publishClientEvent).not.toHaveBeenCalled();
+  });
+});
+
+function createAppInboxIdempotencyHarness() {
+  const queue = new TestResourceInbox();
+  const reader = new InboxQueueReader(queue);
+  const results = new TestResourceInboxResults();
+  const publisher = createPublisher();
+  const runtimeRepository = new FakeRuntimeStateRepository();
+  const database = createAppInboxTestDatabase(queue, results, { runtimeRepository });
+  return {
+    publisher,
+    reader,
+    service: new AppClientInboxService(
+      reader,
+      queue as never,
+      results as never,
+      database,
+      createAutoAuthorizingClientStateService(runtimeRepository, database),
+      'server-12345678',
+    ),
+  };
+}
+
+describe('client mutation service idempotency', () => {
+  it('makes a semantic no-op receipt first-writer-wins', async () => {
+    const runtimeRepository = new FakeRuntimeStateRepository();
+    const service = createClientStateService({
+      runtimeRepository,
+      syncPublisher: createServicePublisher(),
+      now: () => 1_000,
+      serviceId: 'client-service',
+    });
+    await service.upsertPrincipal(CLIENT_MUTATION_SERVICE_SCOPE, 'alice', {
+      username: 'alice',
+      displayName: 'Alice',
+      requestId: 'seed-alice-no-op',
+    });
+    await service.upsertPrincipal(CLIENT_MUTATION_SERVICE_SCOPE, 'alice', {
+      username: 'alice',
+      displayName: 'Alice',
+      requestId: 'alice-no-op',
+    });
+
+    const stored = await new ClientStateRepository(
+      runtimeRepository,
+    ).findIdempotentClientMutationReceipt(toClientPrincipalRef('alice'), 'alice-no-op');
+    expect(stored?.receipt.outcome).toBe('no-op');
+    await expect(
+      service.upsertPrincipal(CLIENT_MUTATION_SERVICE_SCOPE, 'alice', {
+        username: 'alice',
+        displayName: 'Changed',
+        requestId: 'alice-no-op',
+      }),
+    ).rejects.toBeInstanceOf(ClientMutationIdempotencyConflictError);
+  });
+
+  it('rejects the same requestId with different semantic content', async () => {
+    const runtimeRepository = new FakeRuntimeStateRepository();
+    const publisher = createServicePublisher();
+    const service = createClientStateService({
+      runtimeRepository,
+      syncPublisher: publisher,
+      now: () => 1_000,
+      serviceId: 'client-service',
+    });
+    const principalRef = toClientPrincipalRef('alice');
+
+    const first = await service.upsertPrincipal(
+      CLIENT_MUTATION_SERVICE_SCOPE,
+      principalRef.principalId,
+      {
+        username: 'alice',
+        displayName: 'Alice',
+        actorPrincipalId: 'alice',
+        requestId: 'upsert-alice',
+      },
+    );
+    await expect(
+      service.upsertPrincipal(CLIENT_MUTATION_SERVICE_SCOPE, principalRef.principalId, {
+        username: 'alice',
+        displayName: 'Alice with changed payload',
+        actorPrincipalId: 'alice',
+        requestId: 'upsert-alice',
+      }),
+    ).rejects.toBeInstanceOf(ClientMutationIdempotencyConflictError);
+    expect(first.result.right?.event?.eventType).toBe('principal-created');
+
+    const repository = new ClientStateRepository(runtimeRepository);
+    expect((await repository.readSnapshot(principalRef))?.principal.displayName).toBe('Alice');
+    expect((await repository.listEvents(principalRef)).map((event) => event.eventType)).toEqual([
+      'principal-created',
+    ]);
+    expect(publisher.publishClientSnapshot).not.toHaveBeenCalled();
+    expect(publisher.publishClientEvent).not.toHaveBeenCalled();
+  });
+});
+
+describe('client mutation idempotency convergence', () => {
+  it('makes equal request races first-writer-wins and rejects different semantic content', async () => {
+    const runtime = new AggregateBarrierRepository();
+    const request = {
+      username: 'alice',
+      displayName: 'Alice',
+      metadata: { one: 1, two: 2 },
+      requestId: 'same-request',
+    } as const;
+    runtime.armPrincipalReadBarrier(2);
+    const [first, second] = await Promise.all([
+      createService(runtime, 1_000).upsertPrincipal(SCOPE, 'alice', request),
+      createService(runtime, 9_000).upsertPrincipal(SCOPE, 'alice', {
+        requestId: 'same-request',
+        metadata: { two: 2, one: 1 },
+        displayName: 'Alice',
+        username: 'alice',
+      }),
+    ]);
+
+    expect(second.result.right?.event).toEqual(first.result.right?.event);
+    const idempotent = await new ClientStateRepository(runtime).findIdempotentClientMutationReceipt(
+      principalRef('alice'),
+      'same-request',
+    );
+    expect(idempotent?.commandHash).toMatch(/^sha256:[0-9a-f]{64}$/);
+    expect(idempotent?.receipt.commandHash).toBe(idempotent?.commandHash);
+    const records = await outboxFor(runtime, ['same-request']);
+    expect(records).toHaveLength(2);
+    expect(idempotent?.receipt.outboxIds).toEqual(records.map((record) => record.key.resourceId));
+
+    const conflictRuntime = new AggregateBarrierRepository();
+    conflictRuntime.armPrincipalReadBarrier(2);
+    const results = await Promise.allSettled([
+      createService(conflictRuntime, 1_000).upsertPrincipal(SCOPE, 'bob', {
+        username: 'bob',
+        displayName: 'First',
+        requestId: 'different-content',
+      }),
+      createService(conflictRuntime, 1_001).upsertPrincipal(SCOPE, 'bob', {
+        username: 'bob',
+        displayName: 'Second',
+        requestId: 'different-content',
+      }),
+    ]);
+    expect(results.filter((result) => result.status === 'fulfilled')).toHaveLength(1);
+    const rejected = results.find((result) => result.status === 'rejected');
+    expect(rejected).toMatchObject({
+      reason: expect.any(ClientMutationIdempotencyConflictError),
+    });
+    expect(await outboxFor(conflictRuntime, ['different-content'])).toHaveLength(2);
   });
 });
